Refactored Navbar
